+++ src/graphics/text.rs
@@ -54,7 +54,7 @@ impl Tato {
 
 /// Extended from the previous functions to include lowercase letters and additional punctuation.
 #[allow(dead_code)]
-fn char_to_id_ex(ch: char) -> u16 {
+fn char_to_id_ex(ch: char) -> u8 {
     match ch {
         // Tightly packed ASCII chars in their original order can
         // result in fast optimizations by the compiler
